extract on-type formatting into SRP microcrate
