+++ Cargo.toml
@@ -25,6 +25,7 @@ members = [
     "crates/perl-lsp-rename",
     "crates/perl-lsp-protocol",
     "crates/perl-lsp-transport",
+    "crates/perl-lsp-on-type-formatting",
     "crates/perl-lsp-tooling",
     "crates/perl-subprocess-runtime",
     "crates/perl-lsp-formatting-types",
@@ -275,6 +276,7 @@ perl-lsp-capability-map = { path = "crates/perl-lsp-capability-map", version = "
 perl-parser-core = { path = "crates/perl-parser-core", version = "0.10.0" }
 perl-lsp-transport = { path = "crates/perl-lsp-transport", version = "0.10.0" }
 perl-lsp-tooling = { path = "crates/perl-lsp-tooling", version = "0.10.0" }
+perl-lsp-on-type-formatting = { path = "crates/perl-lsp-on-type-formatting", version = "0.10.0" }
 perl-lsp-formatting-types = { path = "crates/perl-lsp-formatting-types", version = "0.10.0" }
 perl-lsp-formatting = { path = "crates/perl-lsp-formatting", version = "0.10.0" }
 perl-tdd-support = { path = "crates/perl-tdd-support", version = "0.10.0" }
