+++ crates/perl-lsp-on-type-formatting/Cargo.toml
@@ -1,0 +1,26 @@
+[package]
+name = "perl-lsp-on-type-formatting"
+version.workspace = true
+edition.workspace = true
+rust-version.workspace = true
+authors.workspace = true
+license.workspace = true
+repository.workspace = true
+homepage.workspace = true
+description = "On-type formatting edit computation for Perl LSP"
+readme = "README.md"
+documentation = "https://docs.rs/perl-lsp-on-type-formatting"
+keywords = ["perl", "lsp", "formatting", "on-type", "ide"]
+categories = ["development-tools", "text-editors"]
+include = [
+  "src/**",
+  "Cargo.toml",
+  "README.md",
+  "LICENSE*"
+]
+
+[dependencies]
+serde_json = "1.0.149"
+
+[lints]
+workspace = true
