+++ crates/perl-lsp-on-type-formatting/README.md
@@ -1,0 +1,5 @@
+# perl-lsp-on-type-formatting
+
+Single-responsibility microcrate for computing LSP `textDocument/onTypeFormatting` edits.
+
+This crate is used by higher-level provider and server crates in the Perl LSP workspace.
